Always serve current version of kit/update script
We previously pointed users to the version of the kit/update script from
the kit main branch if we were running the site locally. This was a
holdover from when the site was also the kit however, and serves no
purpose now, plus it breaks our tests to boot as there is no longer an
update script in the main branch. This commit removes the logic.