Ensure tests install 0.0.1-alpha.3 in test prototype
In commit 5c4410c we tried to make sure that the tests for this repo
would use the use a specific version of the kit, so that the docs tests
are predictable.

This commit adds the desired version to the command line for the kit
install script, because otherwise it will default to the latest
published version.