Make URLs in navigation elements absolute using baseUrl
We can't use relative hrefs in the page layout template, because the
page URL might be nested by one or two levels. Instead, we inject the
express request.baseUrl [[1]] into the Nunjucks environment, which tells
us the part of the page URL that we need to add to the front of the link
URL. (Note, I couldn't find this functionality built in to express, so
I've added it myself, but I'm not wedded to the solution). This should
mean that these links won't need to change if we copy the files to a
different folder for a new version.

Fixes #74.

[1]: https://expressjs.com/en/4x/api.html#req.baseUrl