Add inset text pointing users to v12 docs
For certain pages, add inset text pointing them to the v12 docs, in case
they are still using the old version of the kit.

Note that this commit introduces some tech debt; because we currently
parse pages as Markdown or Nunjucks, but not both, we can't use the
partial on every page in the way that we'd like. I don't have time to
rewrite the renderer now, but we should probably in future.

Co-authored-by: Joe Lanman <joe.lanman@digital.cabinet-office.gov.uk>
Co-authored-by: NoraGDS <57447099+NoraGDS@users.noreply.github.com>