Link to requirements page from getting started page
We want users to be able to follow the 'Get started' link on the
homepage and be led to the command `npx govuk-prototype-kit create`.
This commit makes sure that the 'new user' journey includes that
command, by taking them into the 'getting started' flow that is on the
tutorials and templates page. There is probably a nicer way to do this,
but this is quick and easy for now.