Add redirect from /docs/templates/* to /v12/docs/templates/*
In the v13 docs, we removed the templates from the website. We did this
because the plugin architecture means templates are now decoupled from
the version of the kit a user has. They might have different templates,
or different versions of the same template, to what we show on the
website. As the kit also now has a better built-in way to preview what a
page would look like once created from a template, it seemed like there
was no need to have the templates viewable on the web.

However, what we didn't realise is that the GOV.UK Design System website
links to the template previews on our website, as a way of sharing with
their users what an example use of a component or pattern might look
like. I'm not in a position to figure out if this is how we want
templates to be used in future, or if there is a better way of
fulfilling this use-case, so for now I'm deferring the decision. The
`/docs/templates/*` URLs will work as before, just pointing the user to
the v12 version. In future we might want to install the latest version
of selected plugins and host those templates, or we might want to
negotiate with the Design System team for them to find another way to
get the examples they need.