Use custom template engine function to render Nunjucks pages
When hooking Nunjucks into Express, it changes some core functionality
[[1]]. This makes it hard to use other template engines at the same time
as Nunjucks. This commit changes our code so that Nunjucks is integrated
in a way that is more similar to how Express expects. This will
hopefully mean we can remove some other parts of our our code further
down the line.

[1]: https://github.com/mozilla/nunjucks/blob/fd500902d7c88672470c87170796de52fc0f791a/nunjucks/src/express-app.js