Move handling of .md to server.js
We can remove some duplicated code by moving the logic to remove the
`.md` extension from URLs to the server.js file (and out of the
documentation routes file). Also add some tests for the behaviour of
both this and the similar logic for `.html`, and make sure the
redirected URL has a slash (/) at the end to avoid an unnecessary extra
redirect.

Note: I think we're supposed to be redirecting to the pages with slashes
at the end in all cases, but it's not working? But I don't really want
to try and fix this now, so leaving tests with `skip` for now.