Update README to reflect what this is
This repo contains the docs for the Prototype Kit, not the Prototype Kit itself.

The README is pretty much unchanged from when this was forked from the kit, so update the README to try and reflect what this repo actually contains.

This also means we can remove out of date support info from the README.

I have added a note about the history of this repo which I found confusing.